+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Added: tools to use in the tables
 * Added: config option to disable all tools
 * Added: config option to disable all tool recipes
+* Added: tools made from Thermal Foundation (TF) materials
+* Added: config option to disable all tools made from TF materials
 
 1.1.4
 * Fixed: Tables now work properly with Tinker's tools (#3)